requirements.txt
streamlit
torch
torchvision
opencv-python
Pillow
facenet-pytorch